api/requirements.txt
bcrypt==4.3.0
fastapi[standard]==0.124.4
psycopg2==2.9.11
python-dotenv==1.2.1
python-jose==3.5.0
SQLAlchemy==2.0.45